Apply PR #62 review feedback for user profile page
- Admin badge: move below name, shrink, colour blue
- Personal records: add 333bf.svg event icon, rename NR→LR with
  'Rekord ligi' tooltip on column headers
- Extract FormatWithParens logic to shared SolveFormatHelper;
  Rankings delegates to it
- UserRoundResultDto: add SeasonId for /Rounds deep-link
- UserMatchHistoryDto: add SeasonId and LeagueIdentifier
- Wyniki: Sezon/Kolejka links to /Rounds, Liga column removed,
  A1-A5 columns replaced with inline Ułożenia (best/worst in parens),
  int.MaxValue place rendered as '–'
- Mecze: Sezon/Kolejka links to /Rounds, Liga column added,
  Zobacz styled as btn-primary action button
- Sezony tab renamed to Klasyfikacja; places 1/2/3 show
  gold/silver/bronze trophy icons
- Progresja: Ao5 added as second line on the progression chart
